Project Files/Rice Type Detection/requirements.txt
Flask>=2.3.0
torch>=2.6.0
torchvision>=0.17.0
Pillow>=10.0.0
Werkzeug>=2.3.0
numpy>=1.24.0
requests>=2.31.0 